fix: avoid health checks on windows because they throw `ProcessException`
On windows `SystemResources.init` throws an annoying `ProcessException` with a stacktrace because the command `uname -sm` cannot be executed. This comes from an external package, so it can't be fixed in serverpod.

For developer experience, it would be good to skip the health checks altogether and inform accordingly. I've colored the `stdout.writeln` and defined it as warning to make it clear to the user.